cobA
