--- src/view/LocationView.java
@@ -1,421 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package view;
-
-import controller.CountryController;
-import controller.LocationController;
-import entities.Location;
-import java.util.List;
-import javax.swing.table.DefaultTableModel;
-import org.hibernate.SessionFactory;
-
-/**
- *
- * @author USER
- */
-public final class LocationView extends javax.swing.JInternalFrame {
-
-    private final LocationController locationController;
-    private final CountryController countryController;
-
-    
-    /**
-     * Creates new form LocationView
-     * @param sessionFactory
-     */
-    public LocationView(SessionFactory sessionFactory) {
-        initComponents();
-        locationController = new LocationController(sessionFactory);
-        countryController = new CountryController(sessionFactory);
-        reset();
-    }
-
-    public void edit() {
-        txtLocationId.setEnabled(false);
-        btnSave.setEnabled(true);
-        btnDrop.setEnabled(true);
-        txtCountryName.setEnabled(false);
-    }
-
-    public void reset() {
-//        controller.loadMaxId(txtLocationId);
-        txtKategori.setText("location_id");
-        txtKategori.setVisible(false);
-        txtLocationId.setEnabled(false);
-        txtStreetAddress.setText("");
-        txtPostalCode.setText("");
-        txtCity.setText("");
-        txtStateProvince.setText("");
-        txtCountryName.setText("");
-        txtCountryName.setEnabled(false);
-        btnDrop.setEnabled(false);
-        btnSave.setEnabled(true);
-        bindingLocation((List<Object>) locationController.getAll());
-
-    }
-
-    private void bindingLocation(List<Object> location) {
-        String[] header = {"No", "Location ID", "Street Address", "Postal_Code", "City", "State Province", "Country Name"};
-        String[][] data = new String[location.size()][header.length];
-        int i = 0;
-        for (Object object : location) {
-            Location location1 = (Location) object;
-            data[i][0] = (i + 1) + "";
-            data[i][1] = location1.getLocationId() + "";
-            data[i][2] = location1.getStreetAddress() + "";
-            data[i][3] = location1.getPostalCode() + "";
-            data[i][4] = location1.getCity() + "";
-            data[i][5] = location1.getStateProvince() + "";
-            data[i][6] = location1.getCountryId().getCountryName();
-            i++;
-        }
-        tblLocation.setModel(new DefaultTableModel(data, header));
-        reset();
-    }
-
-    /**
-     * This method is called from within the constructor to initialize the form.
-     * WARNING: Do NOT modify this code. The content of this method is always
-     * regenerated by the Form Editor.
-     */
-    @SuppressWarnings("unchecked")
-    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
-    private void initComponents() {
-
-        jPanel1 = new javax.swing.JPanel();
-        jLabel1 = new javax.swing.JLabel();
-        jLabel2 = new javax.swing.JLabel();
-        jLabel3 = new javax.swing.JLabel();
-        jLabel4 = new javax.swing.JLabel();
-        jLabel5 = new javax.swing.JLabel();
-        jLabel6 = new javax.swing.JLabel();
-        txtLocationId = new javax.swing.JTextField();
-        jScrollPane2 = new javax.swing.JScrollPane();
-        txtStreetAddress = new javax.swing.JTextArea();
-        txtPostalCode = new javax.swing.JTextField();
-        txtStateProvince = new javax.swing.JTextField();
-        txtCountryName = new javax.swing.JTextField();
-        txtCity = new javax.swing.JTextField();
-        cmbKategoriLocation = new javax.swing.JComboBox<>();
-        btnSave = new javax.swing.JButton();
-        btnDrop = new javax.swing.JButton();
-        jLabel7 = new javax.swing.JLabel();
-        jLabel8 = new javax.swing.JLabel();
-        jLabel9 = new javax.swing.JLabel();
-        jLabel10 = new javax.swing.JLabel();
-        jLabel11 = new javax.swing.JLabel();
-        jLabel12 = new javax.swing.JLabel();
-        btnFind = new javax.swing.JButton();
-        txtSearch = new javax.swing.JTextField();
-        cmbKategoriSearch = new javax.swing.JComboBox<>();
-        jScrollPane1 = new javax.swing.JScrollPane();
-        tblLocation = new javax.swing.JTable();
-        txtKategori = new javax.swing.JTextField();
-
-        setClosable(true);
-        setMaximizable(true);
-
-        jPanel1.setBorder(javax.swing.BorderFactory.createTitledBorder("LOCATION DETAIL"));
-
-        jLabel1.setText("Location ID");
-
-        jLabel2.setText("Street Address");
-
-        jLabel3.setText("Postal Code");
-
-        jLabel4.setText("City");
-
-        jLabel5.setText("State Province");
-
-        jLabel6.setText("Country Name");
-
-        txtLocationId.setEnabled(false);
-        txtLocationId.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                txtLocationIdActionPerformed(evt);
-            }
-        });
-
-        txtStreetAddress.setColumns(20);
-        txtStreetAddress.setRows(5);
-        jScrollPane2.setViewportView(txtStreetAddress);
-
-        txtStateProvince.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                txtStateProvinceActionPerformed(evt);
-            }
-        });
-
-        btnSave.setText("SAVE");
-        btnSave.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                btnSaveActionPerformed(evt);
-            }
-        });
-
-        btnDrop.setBackground(new java.awt.Color(255, 51, 51));
-        btnDrop.setText("DROP");
-
-        jLabel7.setText(":");
-
-        jLabel8.setText(":");
-
-        jLabel9.setText(":");
-
-        jLabel10.setText(":");
-
-        jLabel11.setText(":");
-
-        jLabel12.setText(":");
-
-        javax.swing.GroupLayout jPanel1Layout = new javax.swing.GroupLayout(jPanel1);
-        jPanel1.setLayout(jPanel1Layout);
-        jPanel1Layout.setHorizontalGroup(
-            jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanel1Layout.createSequentialGroup()
-                .addContainerGap()
-                .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(jLabel2)
-                    .addComponent(jLabel1, javax.swing.GroupLayout.PREFERRED_SIZE, 86, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jLabel3, javax.swing.GroupLayout.PREFERRED_SIZE, 86, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addGap(12, 12, 12)
-                .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(jLabel9)
-                    .addComponent(jLabel7)
-                    .addComponent(jLabel8))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(txtPostalCode, javax.swing.GroupLayout.PREFERRED_SIZE, 137, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addGroup(jPanel1Layout.createSequentialGroup()
-                        .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                            .addComponent(txtLocationId, javax.swing.GroupLayout.PREFERRED_SIZE, 100, javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addComponent(jScrollPane2, javax.swing.GroupLayout.PREFERRED_SIZE, 274, javax.swing.GroupLayout.PREFERRED_SIZE))
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
-                            .addGroup(jPanel1Layout.createSequentialGroup()
-                                .addComponent(jLabel5, javax.swing.GroupLayout.PREFERRED_SIZE, 86, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                .addGap(11, 11, 11)
-                                .addComponent(jLabel11))
-                            .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
-                                .addGroup(jPanel1Layout.createSequentialGroup()
-                                    .addComponent(jLabel4, javax.swing.GroupLayout.PREFERRED_SIZE, 39, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                    .addGap(58, 58, 58)
-                                    .addComponent(jLabel10))
-                                .addGroup(jPanel1Layout.createSequentialGroup()
-                                    .addComponent(jLabel6, javax.swing.GroupLayout.PREFERRED_SIZE, 86, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                    .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                                    .addComponent(jLabel12))))
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
-                            .addGroup(jPanel1Layout.createSequentialGroup()
-                                .addComponent(btnSave, javax.swing.GroupLayout.PREFERRED_SIZE, 158, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 47, Short.MAX_VALUE)
-                                .addComponent(btnDrop, javax.swing.GroupLayout.PREFERRED_SIZE, 86, javax.swing.GroupLayout.PREFERRED_SIZE))
-                            .addGroup(jPanel1Layout.createSequentialGroup()
-                                .addComponent(txtCountryName, javax.swing.GroupLayout.PREFERRED_SIZE, 132, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                                .addComponent(cmbKategoriLocation, 0, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-                            .addComponent(txtStateProvince, javax.swing.GroupLayout.Alignment.TRAILING)
-                            .addComponent(txtCity, javax.swing.GroupLayout.Alignment.TRAILING))))
-                .addContainerGap(48, Short.MAX_VALUE))
-        );
-        jPanel1Layout.setVerticalGroup(
-            jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanel1Layout.createSequentialGroup()
-                .addGap(15, 15, 15)
-                .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(jLabel1)
-                    .addComponent(txtLocationId, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jLabel4)
-                    .addComponent(txtCity, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jLabel7)
-                    .addComponent(jLabel10))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(jPanel1Layout.createSequentialGroup()
-                        .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                            .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                                .addComponent(jLabel2)
-                                .addComponent(jLabel8))
-                            .addComponent(jScrollPane2, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                        .addGap(12, 12, 12)
-                        .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                            .addComponent(jLabel3)
-                            .addComponent(jLabel9)
-                            .addComponent(txtPostalCode, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)))
-                    .addGroup(jPanel1Layout.createSequentialGroup()
-                        .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                            .addComponent(jLabel5)
-                            .addComponent(txtStateProvince, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addComponent(jLabel11))
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                        .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                            .addComponent(jLabel6)
-                            .addComponent(txtCountryName, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addComponent(cmbKategoriLocation, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addComponent(jLabel12))
-                        .addGap(18, 18, 18)
-                        .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                            .addComponent(btnSave)
-                            .addComponent(btnDrop))))
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-        );
-
-        btnFind.setText("FIND");
-
-        cmbKategoriSearch.setModel(new javax.swing.DefaultComboBoxModel<>(new String[] { "Location Id", "Street Address", "Postal Code", "City", "State Province", "Country Name" }));
-        cmbKategoriSearch.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                cmbKategoriSearchActionPerformed(evt);
-            }
-        });
-
-        tblLocation.setModel(new javax.swing.table.DefaultTableModel(
-            new Object [][] {
-                {},
-                {},
-                {},
-                {}
-            },
-            new String [] {
-
-            }
-        ));
-        tblLocation.addMouseListener(new java.awt.event.MouseAdapter() {
-            public void mouseClicked(java.awt.event.MouseEvent evt) {
-                tblLocationMouseClicked(evt);
-            }
-        });
-        jScrollPane1.setViewportView(tblLocation);
-
-        txtKategori.setEnabled(false);
-
-        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
-        getContentPane().setLayout(layout);
-        layout.setHorizontalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(layout.createSequentialGroup()
-                .addContainerGap()
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(jScrollPane1)
-                    .addComponent(jPanel1, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                    .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, layout.createSequentialGroup()
-                        .addGap(0, 0, Short.MAX_VALUE)
-                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, layout.createSequentialGroup()
-                                .addComponent(cmbKategoriSearch, javax.swing.GroupLayout.PREFERRED_SIZE, 156, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                .addGap(4, 4, 4)
-                                .addComponent(txtSearch, javax.swing.GroupLayout.PREFERRED_SIZE, 175, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                                .addComponent(btnFind, javax.swing.GroupLayout.PREFERRED_SIZE, 80, javax.swing.GroupLayout.PREFERRED_SIZE))
-                            .addComponent(txtKategori, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.PREFERRED_SIZE, 4, javax.swing.GroupLayout.PREFERRED_SIZE))))
-                .addContainerGap())
-        );
-        layout.setVerticalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, layout.createSequentialGroup()
-                .addComponent(txtKategori, javax.swing.GroupLayout.PREFERRED_SIZE, 10, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(2, 2, 2)
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(btnFind)
-                    .addComponent(txtSearch, javax.swing.GroupLayout.PREFERRED_SIZE, 26, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(cmbKategoriSearch, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addGap(18, 18, 18)
-                .addComponent(jScrollPane1, javax.swing.GroupLayout.PREFERRED_SIZE, 236, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                .addComponent(jPanel1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-        );
-
-        pack();
-    }// </editor-fold>//GEN-END:initComponents
-
-    private void txtLocationIdActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_txtLocationIdActionPerformed
-        // TODO add your handling code here:
-    }//GEN-LAST:event_txtLocationIdActionPerformed
-
-    private void txtStateProvinceActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_txtStateProvinceActionPerformed
-        // TODO add your handling code here:
-    }//GEN-LAST:event_txtStateProvinceActionPerformed
-
-    private void cmbKategoriSearchActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_cmbKategoriSearchActionPerformed
-        // TODO add your handling code here:
-        if (cmbKategoriSearch.getSelectedItem().equals("Location Id")) {
-            txtKategori.setText("location_id");
-        } else if (cmbKategoriSearch.getSelectedItem().equals("Street Address")) {
-            txtKategori.setText("street_address");
-        } else if (cmbKategoriSearch.getSelectedItem().equals("Postal Code")) {
-            txtKategori.setText("postal_code");
-        } else if (cmbKategoriSearch.getSelectedItem().equals("City")) {
-            txtKategori.setText("city");
-        } else if (cmbKategoriSearch.getSelectedItem().equals("State Province")) {
-            txtKategori.setText("state_province");
-        } else if (cmbKategoriSearch.getSelectedItem().equals("Country Name")) {
-            txtKategori.setText("country_id");
-        }
-    }//GEN-LAST:event_cmbKategoriSearchActionPerformed
-
-    private void tblLocationMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_tblLocationMouseClicked
-        // TODO add your handling code here:
-        int row = tblLocation.getSelectedRow();
-        txtLocationId.setText(tblLocation.getValueAt(row, 1).toString());
-        txtStreetAddress.setText(tblLocation.getValueAt(row, 2).toString());
-        txtPostalCode.setText(tblLocation.getValueAt(row, 3).toString());
-        txtCity.setText(tblLocation.getValueAt(row, 4).toString());
-        txtStateProvince.setText(tblLocation.getValueAt(row, 5).toString());
-        txtCountryName.setText(tblLocation.getValueAt(row, 6).toString());
-        cmbKategoriLocation.setSelectedItem(tblLocation.getValueAt(row, 6).toString());
-    }//GEN-LAST:event_tblLocationMouseClicked
-
-    private void btnSaveActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnSaveActionPerformed
-        // TODO add your handling code here:
-    }//GEN-LAST:event_btnSaveActionPerformed
-
-
-    // Variables declaration - do not modify//GEN-BEGIN:variables
-    private javax.swing.JButton btnDrop;
-    private javax.swing.JButton btnFind;
-    private javax.swing.JButton btnSave;
-    private javax.swing.JComboBox<String> cmbKategoriLocation;
-    private javax.swing.JComboBox<String> cmbKategoriSearch;
-    private javax.swing.JLabel jLabel1;
-    private javax.swing.JLabel jLabel10;
-    private javax.swing.JLabel jLabel11;
-    private javax.swing.JLabel jLabel12;
-    private javax.swing.JLabel jLabel2;
-    private javax.swing.JLabel jLabel3;
-    private javax.swing.JLabel jLabel4;
-    private javax.swing.JLabel jLabel5;
-    private javax.swing.JLabel jLabel6;
-    private javax.swing.JLabel jLabel7;
-    private javax.swing.JLabel jLabel8;
-    private javax.swing.JLabel jLabel9;
-    private javax.swing.JPanel jPanel1;
-    private javax.swing.JScrollPane jScrollPane1;
-    private javax.swing.JScrollPane jScrollPane2;
-    private javax.swing.JTable tblLocation;
-    private javax.swing.JTextField txtCity;
-    private javax.swing.JTextField txtCountryName;
-    private javax.swing.JTextField txtKategori;
-    private javax.swing.JTextField txtLocationId;
-    private javax.swing.JTextField txtPostalCode;
-    private javax.swing.JTextField txtSearch;
-    private javax.swing.JTextField txtStateProvince;
-    private javax.swing.JTextArea txtStreetAddress;
-    // End of variables declaration//GEN-END:variables
-
-//    private void getSetCmbCountry(){
-//        Vector listCount = new Vector();
-//        List<Object> objects = countryController.getAll();
-//        for (Object object : objects) {
-//            Country country = (Country) object;
-//            listCount.add(country.getCountryId());
-//        }
-//        Collections.sort(listCount  );
-//        set(listCount, cmbKategoriLocation);
-//    }
-    
-}
