+++ src/vieww/CountryView.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package View;
+package vieww;
 
 import controller.CountryController;
 import org.hibernate.SessionFactory;
